Move network port extra DHCP tests
Move the extra DHCP options GET unit test to the standard networking
port package.
