--- openstack/networking/v2/extensions/extradhcpopts/testing/doc.go
@@ -1,2 +1,0 @@
-// extradhcpopts unit tests
-package testing
